split solution in two
Visual Studio only allows one project in a solution to be debugged
at the time. Split the solution in two, so both the player and the
editor can be opened in a separate Visual Studio instance without
file lock conflicts.

Remove outdated Visual Studio 2005 solution and project files while
we're at it.