interpolation: smooth means smoothstep, not cosine
Smoothstep[1] is a nice smooth alternative to cosine that
doesn't require special math functons, and should thus be
suitable on more platforms. The cosine-curve and the
smoothstep are very similar, and this change introduces
only a very small error (always less than 1 per cent)
compared to the old result.

Change the smooth curve definition to match this.

[1]: http://en.wikipedia.org/wiki/Smoothstep