editor: do not pass a string to std::bad_alloc ctor
The C++ standard does not contain a definition of std::bad_alloc
that takes a string, this is something that older versions of
Visual Studio did as an extension. This is no longer available in
Visual Studio 2012, so let's just not pass any string to it.

Thanks to Xeche for noticing.