Bugfix TrackView::editPaste() to not read data from a dead object
mimeData->data(...) returns a QByteArray. Previously this object was
not stored anywhere, but instead a pointer to this array's raw data was
obtained. As the returned QByteArray was destructed immediately after
the pointer was fetched, the clipbuf pointer started to point to
an unused memory area. This caused paste to crash when the memory area
was overwritten by other parts of the program and garbage CopyEntry
objects were read.