player: prevent annoying warning on VS2012 and newer
Visual Studio 2012 and newer seems to have a peculiarity with it's
warning-generation for it's "strdup is reserved, use _strdup
instead"-warning, where it warns even if we use _strdup through a
macro called strdup.

This seems a bit nasty to fix at the core, but it seems we can just
side-step the issue alltogether, by using the NEED_STRDUP macro to
define our own version instead.

This has the side-effect of us using one less CRT function, which
the intro-people may or may not be happy about. But let's let them
weigh in on that if needed.