editor: actively reject pending connections
If we already have a client, we ignore pending connections. This
has a problem; we end up getting more pending connections in our
queue, and since we only process one at the time, we'll never be
able to process the most recent one.

So let's just reject the incomming ones right away. That way,
there's always a one-to-one relationship between the pending
connections, and the processed ones.