editor: refactor TrackView's connected and paused state
We want the MainWindow to keep track of this instead of
trackView, because we're about to get more than one TrackView

But at the same time, the TrackViews need to know if they are
read-only or not.

So let's push a readOnly state down into the TrackView from the
MainWindow when pausing.