player: encode non-portable characters
To prevent non-portable characters to mess up files and whatnot,
let's encode characters not included in POSIX' "Fully portable
filenames" definition. But let's additionally reserve the dash
to denote an encoded byte, so we don't risk generatig the same path
for two different track names.

While we're at it, let's also encode the provided base, so we don't
have invalid characters in that either.