editor: explicitly track active-state
So far, we've been tracking if a track is active or not by counting
how many receivers the keyFrameChanged signal has. However, this only
works as long as only the client subscribes. So instead, let's
explicitly update the active state when the client gets a track, or
disconnect.

This makes it possible to connect other kinds of slots to the signal.