appveyor: do not specify release-name
It seems like it's hard to get Appveyor and Travis CI to agree on
sharing the same release unless I leave the 'release'-option as-is.

So let's just accept this instead of fighting an uphill battle.
This makes things group nicely with the old releases also anyway.

While we're at it, let's drop drafting the release; this only
happens when the tag's already there, so there's no harm in
publishing directly.