editor: rename ClientSocket to SyncClient
SyncClient is a much better name, because the socket-ism of the
client is an implementation detail, not a required part of the
interface of the base-class. So let's generalize this a bit so
we don't leak needless details.

While we're at it, tweak includes of the header a bit for minimal
recompilation.