editor: retain order of in-use tracks when re-creating document
Because the QMap does not retain the order of insertions, if we
were connected to a client and did file -> new, we would re-create
the tracks in an unpredicatble order.

But since all we really store in that map is the index of the track
itself, let's just use a QList instead; it keeps the order nicely.

Sure, getting the index from the track gets a tad slower, but let's
optimize that in the future if it turns out to be a problem.