remove needless GNU-mentions
This project is not a part of the GNU-project, and the old GNU
Rocket joke isn't really that funny. So let's avoid confusion for
outsiders, and just talk about "Rocket" instead of "GNU Rocket".

There's still some cases that have been left due to backwards
compatibility. I might try to write some transitional code at some
later point, but for now this is a step in the right direction.