editor: use QSaveFile if available
In the unlikely event of a power-outage or similar sudden
interruption while the document is writing to disk, it's better to
save to a temp-file and atomically rename that file to the new one
instead of leaving a half-written file on disk.

However, this has several subtleties to it that I don't really want
to have to deal with myself, so let's just use the Qt 5.1 QSaveFile
API for this. This means that older Qt versions will have to live
with the slightly less robust solution for now.