editor: fix crash in debug-build
Taking the address of the first element is invalid when the QVector
is empty, and Qt will trigger an assertion on this in debug-builds.

Luckily, it's completely benign in release-builds, which is what
most people use. But let's make sure we don't annoy those users who
do trigger this, by discarding all their changes!