README: change around the license-badge a bit
Let's point at our own license-file rather than the copy at
opensource.org.

While we're at it, let's have shields.io pick the license from the github
API, instead of specifying. This means there's less things that needs to
be changed if we ever decided to relicence. Not that there's any plans for
that.