editor: update other track-pages when navigating
In 435c884 ("editor: allow navigation while playing"), I forgot that
other pages needs the same update as well. So let's just pipe that
through.

We can safely do it here, because this setter is treated as a
non-navigation row-change, which means that it will not emit an
editRowChanged(int) signal in response.