editor: do paste as single-pass
There's a pretty bad hazard with the previous approach; since
SyncDocument::setKeyFrame doesn't know that we have queued up
deletions inside the paste-region, it will create the wrong command,
leading to the wrong undo-action being performed.

So let's instead set up a map of what we want to paste, and update
the rectangle in one go.