editor: invalidate rest of key if the key is non-step
Pasting (for instance) can add non-step keyframes, and they need
to get their interpolation-mode painted when pasted. So let's make
sure we always invalidate in this case.

The code previously assumed that getPrevKeyFrame would return the
keyframe *before* the current keyframe, but that was a false
assumption. We also couldn't assume there were a keyframe for sure
then either.

So let's clean this up.