editor: store connection objects for some signals/slots
The lifetime of the TrackViews is somehow complicated to keep track of,
and when we're setting the document, we end up trying to disconnect
signals from a TrackView that's currently being deleted.

This is no big deal, as all Qt does is emit a warning and move on. But
it's even nicer to avoid this warning, and we can do that in a pretty
simple way without having to change the code much; by storing the
connection object and disconnecting that instead.

So let's do that. This gets us rid of these three warnings whenever we
load a new document:

QObject::disconnect: No such signal QObject::posChanged(int, int)
QObject::disconnect: No such signal QObject::editRowChanged(int)
QObject::disconnect: No such signal QObject::currValDirty()