sync: add sync_set_val() interface
Add a way for the demo to push values in the opposite direction.

This is intended for more interactive visual manipulation of things
like camera keyframe coordinates.  Within the demo you'd wire up
input events for manipulating camera placement and orientation, and
send the desired values to the editor.  The editor would immediately
update its values for the specified track, and send them back to the
demo as if they had been entered at the editor.