sync: add sync_pause() interface
Add a way for the demo to toggle paused/playing in the editor.

Note this doesn't include a flag payload for communicating the desired
state, unlike the editor->demo direction for PAUSE.

This is intended to be a toggle request for the editor, leaving the
editor authoritative of the playing vs. paused state.  The demo is
a slave to the editor in this regard.