travis: use brew link
brew has a link sub-command that sets up the proper links needed for
qmake. Let's use that instead of setting the QMAKE environment variable.

This shouln't really make a functional difference, except avoid
clobbering the error-code, which made us fail to notice that we failed
to install SDL due to an outdated macOS version being used.