player: enable TCP_NODELAY on Rocket client socket
Rocket's IPC is latency-sensitive, Nagle isn't helping.

This commit only enables it on Darwin and Linux builds.

It's probably trivial to enable it also on WIN32, but since I
don't have such an environment onhand for testing, I've left that
as a TODO/FIXME in device.h for someone on windows to followup.