sync: refactor sync_device into non-blocking connects
This deprecates sync_tcp_connect(), replacing it with
sync_server_setup_tcp() intended to be called just once to
establish the host+port of the server to maintain a connection
with.

Now sync_update() transparently maintains that connection to the
best of its ability, iterating across the addrinfo/hostentry
results of the host resolution done once up front and stored in
the device.

The host+port resolution code is mostly unchanged, just
decomposed to store its iterative state in sync_device instead of
local variables.

A few state variables have been added for tracking if the
sync_device is setup->connecting->connected->ready:

The setup state indicates that sync_server_setup_tcp() has been
called successfully with the resolved address info stores @
sync_device.server

connecting->connected->ready just track the progression of the
connection until established including the greeting etc.  At
which point, sync_update() will start succceeding.

At any point, if d->sock becomes INVALID_SOCKET, it implicitly
clears all of connecting->connected->ready for the cycle to
repeat itself.

Whenever a new connect is initiated, it's performed on the next
address in the address info.  This may not be ideal, it might
make more sense to flag the one we've been connected to as the
one to always use.  But it's mostly harmless.

With this commit, sync_update() should never block the main loop
of the demo.

It would be nice to make sync_update()'s return value describe
the connecting vs. connected state for the demo to show status
overlays or what not.  For now errors are all just -1, they could
be -errno for more context.