sync: re-refactor sync_device non-blocking connects
Based on feedback from @kusma this changes back to an explicit
(re)connect model, with non-blocking connects happening where
supported (win32/*nix) toggled via -DUSE_NONBLOCKING.

When !USE_NONBLOCKING, nbsocket() becomes a noop always
succeeding.  So the attempt to make the socket non-blocking
before connecting just leaves it blocking, and the connect() will
block.  The error handling of the connect() won't go down the
connect-in-progress path unless the socket was non-blocking, so
that's unchanged.

sync_update() maintains the non-blocking connect just like
before, finishing it if possible, as well as consuming the
greeting from the server in both blocking/non-blocking connect
scenarios, after which the device is consiered "ready" and
updates will be performed.

But if disconnected, sync_update() no longer automatically starts
a reconnect.  Instead it returns an error, and the caller must
initiate a connect via the resurrected sync_connect() (its
signature is different at least).

The removal of automatic reconnects also means the initial
connect must now be performed explicitly via sync_connect() as
well.

But as before, the server name resolution/setup is performed in a
separate step from initiating the connect.  The resolution is
still done in a plain blocking manner, storing the results in
struct sync_device.  sync_connect() will iterate through the
resolution results transparently on every connection attempt.